+++ server/cartaro/model/time_off/personal.py
@@ -63,11 +63,11 @@ class Personal(TimeOff):
     def __loadWorkDays(self):
         # Beginning of year
         startDate = arrow.get(F"{self.year}-01-01")
-        
+
         endDate = arrow.now()
         if self.year != endDate.year:
             endDate = arrow.get(F"{self.year}-12-31")
-        
+
         days = WorkDay.find(
             op="and",
             sort_by="date",
@@ -86,23 +86,30 @@ class Personal(TimeOff):
         return used
 
     def available(self):
-        available = self.starting_balance + self.accrued_ytd()
-        return available
-    
+        return self.starting_balance + self.accrued_ytd()
+
     def balance(self):
-        balance = self.available() - self.used
-        return balance
+        return self.available() - self.used
 
     def accrued_ytd(self):
         accrued = 0.0
         if self.accrual:
             now = arrow.now()
             if self.year < now.year:
-                avail_months = 12    
+                avail_months = 12
             elif self.year == now.year:
                 avail_months = arrow.now().month
             elif self.year > now.year:
                 avail_months = 0
 
             accrued = (avail_months / self.accrual_period) * self.accrual_rate
+
+            # Can't accrue more than accrual.cap
+            # Accrual *stops* when reach cap
+            # Adjust for that
+            if self.accrual_cap is not None:
+                total_avail = (self.starting_balance + accrued) - self.used
+                if total_avail >= self.accrual_cap:
+                    accrued = accrued - (total_avail - self.accrual_cap)
+
         return accrued