+++ src/renderer/models/PTO.js
@@ -7,6 +7,10 @@ class PTO extends Resource {
   static RESOURCE_NAME = 'time_off/personal'
   static DEFAULT_ICON = 'mdi-timetable'
 
+  startingBalance () {
+    return parseFloat(this.starting_balance)
+  }
+
   icon () {
     let icon = Icon.superSearch(this.type)
 
@@ -39,31 +43,34 @@ class PTO extends Resource {
         availMonths = 0
       }
       accrued = (availMonths / this.accrual.period) * this.accrual.rate
+
+      // Can't accrue more than accrual.cap
+      // Accrual *stops* when reach cap
+      // Adjust for that
+      if (this.accrual.cap) {
+        const totalAvail = (this.startingBalance() + accrued) - this.used
+        if (totalAvail >= this.accrual.cap) {
+          accrued = accrued - (totalAvail - this.accrual.cap)
+        }
+      }
     }
     return accrued
   }
 
   cap_reached () {
     let over = false
-    if (this.accrual && this.balance() >= this.accrual.cap) {
-      console.log(`${this.type} is over: ${this.balance()} >= ${this.accrual.cap}`)
-      over = true
+
+    if (this.accrual && this.accrual.cap) {
+      if (this.balance() >= this.accrual.cap) {
+        over = true
+      }
     }
+
     return over
   }
 
-  availableYTD () {
-    return parseFloat(this.starting_balance) + this.accruedYTD()
-  }
-
   available () {
-    let available = this.availableYTD()
-
-    if (this.accrual && available > this.accrual.cap) {
-      available = this.accrual.cap
-    }
-
-    return available
+    return this.startingBalance() + this.accruedYTD()
   }
 
   balance () {